+++ forui/lib/widgets/sidebar.dart
@@ -1,0 +1,11 @@
+/// {@category Widgets}
+///
+/// A sidebar widget that usually resides on the side of the screen for navigation.
+///
+/// See:
+/// * https://forui.dev/docs/layout/sidebar for working examples of a sidebar.
+library forui.widgets.sidebar;
+
+export '../src/widgets/sidebar/sidebar.dart';
+export '../src/widgets/sidebar/sidebar_group.dart';
+export '../src/widgets/sidebar/sidebar_item.dart';