+++ forui/CHANGELOG.md
@@ -6,6 +6,15 @@
 
 * Fix style suggestions always displaying actual style name instead of alias.
 
+
+### `FSidebar` (new)
+A sidebar widget that usually resides on the side of the screen for navigation.
+
+* Add `FSidebar`.
+* Add `FSidebarGroup`.
+* Add `FSidebarItem`.
+
+
 ### `FTabs`
 
 * Add `FTabs.onPress`.